made minor edits to make windows compatible

+++ HPP/simulation.hpp
@@ -20,7 +20,7 @@ void Simulate(const list<string> countryNames, list<int> population) {
     list<Country> infectedCountries;
     list<Country> overtaken;
     long int totalInfected = 1;
-    const long int WORLD_POPULATION = 7808449172;
+    const long long int WORLD_POPULATION = 7808449172;
     int day = 1;
     int lastID = 1;
     bool skip = false;
